Adding MD5 of the app/config/routes.php to $key
This removes the need to clear the cache when updating custom routes, fully automating the Helper modification

Note that if the application injects routes into the router - from the Database for example - then updating these injected routes will still require clearing the cache. An alternative method would be to define a Cache configuration on the fly in this file and clear only that configuration when the injected routes are updated.